move out additional util methods, use shorthands for api
